- wip
- wip
- wip
- wip
- wip
- wip
- nuget update
- removed .sql file
- wip
- wip
- nuget
- nuget
- SystemEndpoints
- wip
- build trigger
- Merge branch 'main' into features/jobs
- wip
- added persistent LastExecuted to JobBase
- wip
- wip
- added safeWhere extension
- added jobs http endpoints (get/start)
- add more job properties
- more job properties
- add more job properties (u.a corrid)
- pipe job endpoint corrid to job ctx
- wip
- dotnet update
- better Vulnerability scan
- wip
- moved IStartupTask to abstractions + add new quartz db seeder
- wip
- wip
- quartz sqlserver config
- nuget updates + serilog 4
- wip
- sql prefix
- wip
- refactor primary ctors
- debug level for db loginterceptor
- namespaces
- fix weatherforecast app
- add new dotnet 8 blazor app (weatherforecast)
- more primary ctor refactorings
- cleanup usings
- dinnerfiesta more primary ctors
- Update null checks syntax across multiple files
- nuget updates
- dinnerfiesta rules/specs partials
- wip
- improve MigrationsHostedService
- dbmigratorservices added EnsureTruncated
- wip
- wip
- wip
- wip
- wip
- improve DatabaseCreatorService
- Refactor table database truncation logic
- wip

+++ src/Infrastructure.EntityFramework/Database/DatabaseCreatorService.cs
@@ -9,6 +9,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using BridgingIT.DevKit.Common;
+using BridgingIT.DevKit.Infrastructure.EntityFramework.Database;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Storage;
@@ -43,50 +44,80 @@ public class DatabaseCreatorService<TContext> : IHostedService
             return;
         }
 
+        var contextName = typeof(TContext).Name;
+
         if (this.options.StartupDelay.TotalMilliseconds > 0)
         {
-            this.logger.LogDebug("{LogKey} database creator startup delayed (context={DbContextType})", Constants.LogKey, typeof(TContext).Name);
+            this.logger.LogDebug("{LogKey} database creator startup delayed (context={DbContextType})", Constants.LogKey, contextName);
 
             await Task.Delay(this.options.StartupDelay, cancellationToken).AnyContext();
         }
 
         try
         {
+            //this.logger.LogDebug("{LogKey} database creator initializing (context={DbContextType})", Constants.LogKey, contextName);
             using var scope = this.serviceProvider.CreateScope();
             var context = scope.ServiceProvider.GetRequiredService<TContext>();
-            if (context.Database.ProviderName != "Microsoft.EntityFrameworkCore.InMemory")
+            if (!this.IsInMemoryContext(context))
             {
-                this.logger.LogInformation("{LogKey} database creator started (context={DbContextType})", Constants.LogKey, typeof(TContext).Name);
+                this.logger.LogDebug("{LogKey} database creator started (context={DbContextType}, provider={EntityFrameworkCoreProvider})", Constants.LogKey, contextName, context.Database.ProviderName);
 
-                if (this.options.EnsureDeleted)
+                var exists = await context.Database.CanConnectAsync(cancellationToken);
+                if (exists && this.options.EnsureDeleted)
                 {
+                    this.logger.LogDebug("{LogKey} database creator delete tables (context={DbContextType})", Constants.LogKey, contextName);
                     await context.Database.EnsureDeletedAsync(cancellationToken).AnyContext();
+                    exists = false;
                 }
 
-                await context.Database.EnsureCreatedAsync(cancellationToken).AnyContext();
-
-                try
+                if (exists && this.options.EnsureTruncated)
                 {
-                    // creates tables of additional DbContexts
-                    var databaseCreator = context.Database.GetService<IDatabaseCreator>() as RelationalDatabaseCreator;
-                    databaseCreator?.CreateTables();
-                }
-                catch
-                {
-                    // ignore, tables are already created
+                    this.logger.LogDebug("{LogKey} database creator truncate tables (context={DbContextType})", Constants.LogKey, contextName);
+                    if (context.Database.ProviderName == "Microsoft.EntityFrameworkCore.SqlServer")
+                    {
+                        await context.Database.ExecuteSqlRawAsync(
+                            SqlStatements.SqlServer.TruncateAllTables(this.options.EnsureTruncatedIgnoreTables), cancellationToken).AnyContext();
+                    }
+                    else if (context.Database.ProviderName == "Microsoft.EntityFrameworkCore.Sqlite")
+                    {
+                        await context.Database.ExecuteSqlRawAsync(
+                            SqlStatements.Sqlite.TruncateAllTables(this.options.EnsureTruncatedIgnoreTables), cancellationToken).AnyContext();
+                    }
+                    else
+                    {
+                        throw new ArgumentException($"Database provider '{context.Database.ProviderName}' does not supported truncating tables.");
+                    }
                 }
 
-                this.logger.LogInformation("{LogKey} database creator finished (context={DbContextType})", Constants.LogKey, typeof(TContext).Name);
+                //await context.Database.EnsureCreatedAsync(cancellationToken).AnyContext();
+
+                // alternative way for EnsureCreatedAsync
+                // also creates tables of additional DbContexts which are housed in the same database
+                var databaseCreator = context.Database.GetService<IDatabaseCreator>() as RelationalDatabaseCreator;
+                if (!exists)
+                {
+                    await databaseCreator?.CreateAsync(cancellationToken);
+                    exists = true;
+                }
+
+                await databaseCreator?.CreateTablesAsync(cancellationToken);
+
+                this.logger.LogInformation("{LogKey} database creator finished (context={DbContextType}, dbexists({DbExists}), provider={EntityFrameworkCoreProvider})", Constants.LogKey, contextName, exists, context.Database.ProviderName);
             }
         }
         catch (Exception ex)
         {
-            this.logger.LogError(ex, "{LogKey} database creator failed: {ErrorMessage} (context={DbContextType})", Constants.LogKey, ex.Message, typeof(TContext).Name);
+            this.logger.LogError(ex, "{LogKey} database creator failed: {ErrorMessage} (context={DbContextType})", Constants.LogKey, ex.Message, contextName);
         }
     }
 
     public Task StopAsync(CancellationToken cancellationToken)
     {
         return Task.CompletedTask;
+    }
+
+    private bool IsInMemoryContext(TContext context)
+    {
+        return context.Database.ProviderName == "Microsoft.EntityFrameworkCore.InMemory";
     }
 }
